Moved round method to machineHumiData class
